adds examples for optimization-based trajectory generation
